[sources] Turn on 'Code folding' by default.
This toggles the 'Code folding' setting to be on by default, matching
what VScode does. Otherwise the feature is barely discoverable, despite
interest from developers in using it. We made the decision to change the
default, since even when the developer doesn't actively use it, it only
consumes very little horizontal space and doesn't draw unnecessary
attention to it.

Fixed: chromium:1459193
Change-Id: I7f5bcf8b35e88e9aa94e33872611f5db382ab488