Extract lines from CPU profile coming from trace
This property enables the feature to provide line level execution time
of a function in the sources panel.

Unlike the CDP.Profiler.Profile type, the profile coming in trace events contains the lines data as part of the profile itself, making its type shape different from that of CDP. For this reason a new type is introduced in the CPU Profile parsing library, which inludes and optional `lines` property.

When using this type explicitly, TS caught some discrepancies around the ProfileNode.callFrame type from CDP and the type we get from a trace, this CL also solves these issues.

Bug: 1457230
Change-Id: I8db7ae63012c465c2e2bf6366ab564488b4c77ca