Simplify and tighten render coordinator job queues
In our codebase, we only provide labels for coordinator write jobs,
but never for reads and scroll jobs. We also always return void
for write and scroll jobs.

This patch makes all of that explicit in the implementation and
in the types.
* Read and scroll jobs never skip callbacks.
* There is no unsafe type casting when we enqueue jobs.

We could go even further and remove the unused labelled versions
of the scroll and read methods, but they are useful for debugging,
so let's keep them around.

Bug: none
Change-Id: I968a3b8b4a8569e60d83903fc8c5716172789195