Remove nesting popover and add nesting selectors directly
This CL removes the original nesting popover and adds the nesting
selectors directly above the current rule selector, along with other
at-rules, in the order that's authored.

The effort to update specificity popover info will come in a separate
CL, after the backend is updated as well.

Screenshot: https://imgur.com/a/f3LvmgE

Bug: 1172985
Change-Id: I9ba7ba257735d166018c6942aef4a8a644f0c162