Migrate SDK.TracingModel and TracingManager to models/trace
To avoid circular dependencies when we need to use the SDK module in the
new engine, this CL moves the TracingModel and Manager into models/trace
and updates references across the codebase.

Note that some unit test files (TracingModel_test and
TraceSDKServices_test) have not been moved and need to be relocated now
that their implementation is in models/trace. However this CL is already
large enough so I will do that in a follow-up CL.

Bug: 1462575
Change-Id: Iee0cd7a697fd2124c325bc79901fb5ce0c4c877f