[scopes] Cache variable mapping on parsed scope instead of 'pause' scope
Currently we cache the variable name mapping on the scope chain entry
object we get per "Debugger.paused" event. This means we re-calculate
the variable mapping on every step for every scope in the scope chain.

This CL shuffles around some existing code so we can cache on the
parsed "ScopeTreeNode" instead. This comes with the additional
overhead of finding the parsed scope based on the source position of
the "pause" scope, but local testing has shown that this is
reasonably fast and we can speed it up even further by using a binary
search.

R=jarin@chromium.org

Bug: 1444349
Change-Id: I7adad51a06aac5959ca25203d29161b9e3f690c6