[RPP] Cache frame by process id
Up until now we were tracking which process ids are present in a given
frame, but there was no straight forward way to know which frame owns
a given process (without having to traverse the frames to process map).

For practicality, this CL adds a map that tracks process to frame. This
way we will be able to match data from the renderer handler (groupped
by process) with URL (obtained from the frame).

Bug: 1428024
Change-Id: I442c92171b3fefe369bcba0512fe0c5c99a85030