Reset trace processor in Network Panel
This fixes a bug where the trace processor will error on the 2nd (and
subsequent) attempts at recording because it was not reset properly.

I also as a drive-by renamed the method to resetProcessor, as reset()
implies that it clears out all the model data and any historical traces,
which is incorrect. It simply resets the processor so it is ready to be
used again.

Bug: none
Change-Id: I8cd8cc49f53b4892724fc7eed6b271de049da7d6