[RPP] Cache frame by process id
Up until now we were tracking which process ids are present in a given
frame, but there was no straight forward way to know which frames are
present in a given process (without having to traverse the frames to processes map).

For practicality, this CL adds a map that tracks process to frames. This way we will be able to match data from the renderer handler (grouped by
process) with URL (obtained from the frame).

Bug: 1428024
Change-Id: I442c92171b3fefe369bcba0512fe0c5c99a85030
f