[breakpoints] Substitute variables in breakpoint conditions
This CL extends the 'evaluateExpressionsWithSourceMaps' to
conditional breakpoints. To make this work, we need to add
a new method 'allVariablesAtPosition' in NameResolver. This
is because the existing 'allVariablesInCallFrame' is not
sufficient for conditional breakpoints as those can be
set independent of where the page is paused.

'allVariablesAtPostion' mostly mirrors 'allVariablesInCallFrame'.

Note that 'Breakpoint#backendCondition()' now has two overloads:

  1) If no location is passed in, we don't substitute variables.
     In this case there is no async work so we return no promise.

  2) If a location is passed in and the experiment is enabled we
     return a promise since variable substitution can take some
     time.

We implemented 'backendCondition' this way because otherwise we'd
have to propagate async/await up the call-chain and in some places
we want to keep things synchronous (e.g. initalizing the
BreakpointManager).

R=jarin@chromium.org

Bug: 1444349
Change-Id: I59af6fbfe39ce08daff1f4662342542e0199bd12