Export parsed CPU profile in samples handler
At the moment we export the raw profile we get from tracing along with
the profile calls built from it. This CL adds the parsed profile to the
exported data, which consists of the CPUProfileDataModel instance
created for every CPU Profile in a trace.

This will be used to integrate trace event data with samples. In that
integration step we will create profile calls, which are different from
the ones created in the SamplesHandler at the moment because the latter
are created without taking trace events into account.

Finally, as a drive by change, this also caches the exported data by
thread id, instead of profile id. This is because the profile id is
unknown outside the context of the samples handler and as such we need
the thread id each profile belongs to when integrating with the
renderer handler.

Bug: 1449191
Change-Id: Icd66813b73ea4e7517ad4673fa9e4a340d1a534f