Fix bad bounds from overview component
This change was introduced in crrev.com/c/4580489 however by setting the
bounds to 0 and Infinity it causes a bug in the Performance Panel where
the user dragging the handles can drag beyond the limits of the trace.

This CL updates the times to not be set to 0 or Infinity which fixes the
bug. It also does not break the new intended behaviour introduced by the
CL above, as that CL also contains code in the NetworkPanel to catch
this case and update it correctly.

Fixed: 1464269
Change-Id: Ifaa7b6bc75aead2bca5b13e051d2f5846c0d771b