Add ui/legacy/tokens.css to component server
This way we ensure colors loaded from this file are available in the
component examples and screenshot tests, since it seems that after
https://crrev.com/c/4650363, some colors for icons were no longer
available in the component server (see screenshots updates).

This CL adds the missing stylesheet to the server and adds it to a
reusable array so that stylesheet paths don't have to be hardcoded
multiple times.

Bug: none
Change-Id: I8adcfbdd4966706a9b6f15d0947b4ef0ac0e96f4