Introduce SamplesIntegrator helper to trace model
Its purpose is to aid in the integration of samples into trace events.
It creates profile calls with durations using the CPU profile samples
and the trace events.

Its logic is analogous to the OPP:
Tracking the state of the stack as the following events happen (in
order):
 1. A sample was done.
 2. A trace event started.
 3. A trace event ended.
Depending on the event and on the data that's coming with it the stack
is updated by adding or removing JS calls to it and updating the
duration of the calls in the tracking stack.

The helper isn't used yet, that will be done as a follow up.

Bug: 1449191
Change-Id: Ic7a3d05fa4c16ddd2b60ffc4b821404b5f29d58a