Ensure we resume all targets when performance tracing errors
If there is an error in the actual tracing, the TimelineController
should handle it and resume all targets. But because resumeAllTargets()
is safe to call multiple times (it no-ops if nothing is suspended), we
can also call it again if the try{} block in the TimelinePanel that
starts recording errors.

Fixed: 1462539
Change-Id: Ic52f72617447b955d02e512d18edade76293676e