Use primaryPageTarget in Perf Panel
This CL fixes the two attached bugs where we were trying to calculate
the URL to trace and it was all going wrong because the controller's
main target was incorrect, and we need to use the primaryPageTarget
method.

As a follow-up CL, we will explore adding test coverage to this flow to
ensure it cannot break like this.

Fixed: 1464725,1464831
Change-Id: Ifbe784fb39545696a10ade28b63850ebfa156a49