Create TimelineMiniMap widget
This CL does not change anything for users but does introduce a nice
container for the minimap which can be rendered in the components
server. It also means we can house the minimap styling in a specific
place, not in timelinePanel.ts, and therefore it works just fine when
rendered in the component server.

Wrapping this code in this widget will make it easier to refactor as
work continues to migrate it away from the legacy data model and onto
the new engine.

Bug: 1464206
Change-Id: Iaffdde0825aed69496f0bb047691046873a81f5c