Fix errors when running component server index
Passing an array to `path.join` is not supported, it takes unlimited
arguments rather than an array of one.

This was introduced in crrev.com/c/4678757, but I think the bots were
happy because they do not use the index page and instead navigate
directly to examples. This error only triggered on the index page.

As a drive-by, I also bumped the font size as otherwise they use the
default, rather small, DevTools font size.

Bug: none
Change-Id: I580c454d62e3ce66065bdb21c2fc7b0349825637