Modify SamplesIntegrator to be executed in one step
After working on a prototype to use the SamplesIntegrator in the
renderer handler I noticed it is more convenient to call the samples
integrator one time with the trace events to build the profile calls.

The logic we had before assumed we would call the integrator as we
built the tree derived from trace events in the RendererHandler (the
step which determines the parent-child relationships among events). The
integrator would be called each time a trace event started and ended or
when a sample happened. I now notice two problems with this approach:

1. Building profile calls from trace events and samples at the same time
   as we build the tree from trace events and profile calls results in
   making code that's hard to understard even harder. This has the
   caveat that doing these in two different steps would imply going
   through all events twice, so there's a potential hit in the trace
   engine performance, however:
2. We are building the tree lazily anyways when we create the data for
   the trees in the details pane (top down and bottom up trees). It
   isn't necessarily the case that we need to build it preemptively at
   trace parsing in the RendererHandler.

With this in place, the renderer handler will be changed to call the
samples integrator to get profile calls and export the result of
merging these with the trace events it tracks.

Bug: 1428024
Change-Id: I476abd9195edd6b89e22f39518cf73f8358ac4b1