Make the RendererHandler compatible with ProfileCalls
This is a preparation to start integrating profile calls in the
RendererHandler. We want to export the merge between trace events and
profile calls for each thread, so this CL changes the name of this
property from "events" to entries. Other references to events are also
changed to refer to entries.

This CL also moves some types to the Types/TraceEvents.ts file for
consistency.

Bug: 1428024
Change-Id: I9ef4112386115d5f00a9707e117c17b60ff8e321