[scopes] Mark shadowed variables as 'unavailable'
This CL changes 'allVariablsInCallFrame' and 'allVariablesAtPosition'
to mark original names as 'null' when the underlying compiled name
was shadowed.

Otherwise we'd use the wrong variable after substitution.

Note that we do not yet report failures to the UI and all use-sites
currently fall back to using the original expression.

Drive-by: Type worker params as `unknown` internally. Ideally we'd
have a helper type for "structured cloneable" going forward.

R=jarin@chromium.org

Bug: 1444349
Change-Id: Id0040eacaf8076fd1ee62f9a08edfe26bc12dc49