Add eslint rule for detecting TraceEngine test timeouts
We need to ensure that each test that uses the trace engine is defined
using the function() keyword rather than an arrow function so that the
timeout is correctly set. This CL adds a rule to enforce and autofix
that.

Note that this CL does not enable the rule, I will do that in a
follow-up.

Bug: none
Change-Id: I33c1d2ea4bf8e639fd6b58dcabab8655b1ff222c
No-Tree-Checks: True
No-Tree-Checks: True