[RPP] Support  multiple URLs by renderer process
Before this CL, the MetaHandler supported the possibility to assign
multiple processes to each frame. However it was not contemplating the
possibility that a given process might have different URLs. This is an
actual possibility and it is supported by the OPP's model. I noted this
discrepancy after seeing the same track being rendered with different
URLs when appended by the OPP and by the new engine.

This changes the data structure in the MetaHandler that tracks processes
in frames so that multiple windows are tracked for each process, where a
new window is pushed if we detect that the process URL changed. The
RendererHandler is also changed so that the URL assigned to a process
corresponds to the first rendered frame's URL (the process URL is the
one we use in the process' track).

Bug: 1428024
Change-Id: Iac8a3280b41d68144c57be93c17407474cba4750