Fix manifest screenshot aspect ratio
The screenshots section in the manifest pane is showing the wrong aspect ratio of the image. The screenshots currently as a fixed 1:1 aspect ratio.

However, since most of the screenshots are not in 1:1 aspect ratio (ie: 16:9), we should add "object-fit: contain" to the image in order to make sure the image is showing in the original aspect ratio.

Bug: 1468623
Change-Id: I4260dbbe4b7f21743995f0139c7aa50907678804