[TextEditor] Restore default CodeMirror Ctrl-Arrow behavior.
With https://crrev.com/c/4517865 we not only changed the Alt-Arrow
behavior for the text editor (to jump based on syntax), but also altered
the Ctrl-Arrow behavior to jump between subwords. This however does not
align with VSCode's behavior for Ctrl-Arrow, which basically jump based
on words / non-words.

With this change we now match VSCode's behavior for Ctrl-Arrow
(Alt-Arrow on macOS).

Fixed: chromium:1468208
Bug: chromium:1241848, chromium:1299799
Change-Id: Ifcc7ab18affb7135048587844ad624ac5b3450e3