Fix `any` types from handlers
We currently want to mark these handlers as optional because they are
only used when running the component server and not when running the
actual engine in production - is is to enable us to migrate the main
thread track incrementally.

When we pass in optional properties to the EnabledHandlerDataWithMeta
type, something goes wrong, and the types for Renderer and Samples get
marked as `any`.

To fix this, we can just hard code these in as optional types. This will
only be a temporary bit of code, as soon these handlers will be properly
enabled and non-optional, and we can remove this hardcoded line.

Bug: none
Change-Id: I56b4dd5e5ef95df27b5b943e354a3999cf3f816e