[Protocol Monitor] Add custom editor for object parameters with no predefined keys
For very rare object parameters, their keys are defined in the CDP documentation.
For this very specific type of parameters, a custom editor is displayed to the user, allowing him to directly edit both the keys are the values for the nested parameters of this object.
The nested parameters will be treated only as string parameters.

Bug: 1464495
Change-Id: Idd6cd7f21db9387f7b4c098eadc430272e184ea9