[Animations] Track metrics for animations panel actions
Here are the added stuff with descriptions.

Actions:
* AnimationGroupsCleared: when the clear button is pressed.
* AnimationsPaused: when the pause button is pressed.
* AnimationsResumed: when the resume button is pressed.
* AnimatedNodeDescriptionClicked: when the user explicitly wanted
to jump to the animated node from Animations panel. The reason I
have named the action this way is, there is another way an animated
node might be revealed: when the user is playing with the timeline
points and the animations panel is opened as a drawer. We want to
explicitly know whether people explicitly jump to the node definition
or not.
* AnimationGroupScrubbed: when the scrubber moved and its move is ended.
* AnimationGroupReplayed: when the animation group is replayed.

Enumerated histograms:
* AnimationPlaybackRateChanged: user changed the playback rate to some other value. This also captures which value it is changed to.
* AnimationPointDragged: user moved a point in the animation timeline to
edit the animation duration or delay.

Bug: 1470737
Change-Id: I3b2b7c037b12a5e3ac3d5a319a95cd174f26e217