Identify CompositorTileWorker threads in RendererHandler
These threads are the ones that are shown as Rasterizer tracks in the
panel, so we should identify them in the handler in order to then be
able to easily generate tracks for them.

We could do this in the UI layer based on going through all the
processes and threads, but identifying these in the handler feels right
to avoid the UI having to do too much additional processing.

Bug: 1471399
Change-Id: Ic64b215bedf651475c421b6e633e5ca61ad10833