Rename navStartTimes()
The name `navStartTimes` in the TimelineFlameChartDataProvider type was
confusing as it returned a map of NavigationID=>TraceEvent. But where we
used this function we didn't care about the navigationID or the map
structure. So in this CL I have renamed the method and changed it to
return an array to make it easier to work with.

Bug: none
Change-Id: Ia5855144839a2c8778b472645d7627ea7a2bb0f3