[search] Preserve search results in 'findFilesMatchingSearchRequest'
This CL changes the 'findFilesMatchingSearchRequest' return type from
a list of UISourceCodes to a Map keyed on UISourceCodes.
The ContentProviderBasedProject already does a prelminary search in
the backend and throws a way the search result.

THis CL is a first step to re-use the search result by returning it.
The "file based project" implementation unconditionally omits the
search result.

R=dsv@chromium.org

Bug: 1468875
Change-Id: Id46df819973e1caeb0dcd2b605012de35f60c145