Update UIUtils to not worry about legacy events for nav adjustment
Now that both the network and flame chart data provider return
navigations from the new engine, we can simplify the helper in UIUtils
to only worry about events from the new engine.

This is step one of a few CLs that will remove navigation start event
logic from the old engine entirely as we can rely exclusively on the new
engine for it.

Bug: 1472229
Change-Id: I64067f1e39742c17284ccbff8033044508cc4aae