Remove `setNavStart` method from minimap and use new engine
This CL updates the TimelineMiniMap to use the new engine for its
navigation events. We remove the explicit method that was used to set
the navigation start events, and instead calculate them via the new
engine.

I also renamed the MiniMap `updateControls` method to `setData` to
better reflect what it does.

After this CL the old engine's tracking of navigation start events is
entirely unused; I will therefore remove it but do so in a follow-up to
avoid this CL getting too large.

Bug: 1464206, 1472229
Change-Id: If5379434b8ea91504b3783b97f92b5fc7b4c127f