PreloadingDevTools: do not truncate tooltips on urls
How it works:
Previously the Cell.title field was used to display tooltips.
And it is not set, Cell.value (the url) will be truncated to the first
20 characters and be displayed.

So we can set the Cell's title field to the complete url to avoid
truncation.

Before: See the issue's description
After: https://crbug.com/1470614#c4 and https://crbug.com/1470614#c7

Bug: 1470614
Change-Id: Iaf4fe9083cb4b399b87eb57c4b1468facf88b245