Fix network log stick-to-bottom
Height attributes are rounded to the nearest integer. This means that
arithmetics with these values can accumulate errors. The previous fix
for crbug/1349531 did not account for this potential error.

Instead of relying on UI measurements, this fix uses the logic of
whether the last log entry is visible. That seems a lot more robust.

R=jacktfranklin@chromium.org

Bug:1472300
Change-Id: I3121264dc41a0acda2d5cf396ebbb92f1596c091