Introduce first version of StackManipulator
This CL lands the very first version of the StackManipulator which is
intended to provide support for various user actions on a flame chart,
including collapsing stacks, merging functions, collapsing recursion,
and so on.

There is plenty more functionality we want to implement (see attached
bug and subsequent design doc); this CL just lands the initial version
and a few tests of the basic functionality that has been implemented.

Bug: 1469887
Change-Id: Icbfb2f59d9788d70fbba814ccc9fee6ccd1d2fba