[scopes] Replace startLocation/endLocation with 'range'
This CL slightly improves the types for scope start/end. Instead of
separate locations we use LocationRange. This means the type system
guarantees that either both are present or neither.

We also move the check that the scriptId of both start and end matches
into scope construction and remove it in the NameResolver. Note that
V8 guarantees this property by construction but other backend
implementations exist.

R=jarin@chromium.org

Bug: None
Change-Id: If0de5a2ab345b4643407026b133b501d8118fc83