Fix MetaHandler calculating incorrect trace start time
In the situation where there is no TracingStartedInBrowser event, the
MetaHandler would incorrectly return -1 as the start time. The fix is to
check we found a TracingStartedInBrowser event, otherwise we fallback to
the earliest valid start time we found as we were parsing the trace.

This enables us to rely on these bounds for web usage and also Node/CPU
Profiling usage.

Bug: none
Change-Id: I431c88b336d7a6c8df0cd579221c5d71b92b0f62