[editor] Use different CM scroll strategy for ranges
The current 'revealPosition' logic uses 'center' as the scrolling
strategy. This is to provide context around the revealed position as
its a single point. With 'nearest' the single point is very likely
revealed at the very edge of the editor, with the actual interesting
bits offscreen.

If the selection is only a single point, this makes sense. But if the
caller provides an actual range, we can use the default 'nearest' as
CM ensures that the whole range will be scrolled into view.

We'll use this for revealing search matches.

R=jarin@chromium.org

Bug: 1468875
Change-Id: I14af66e16d311a82c17bf256c44d6eceb98725b3