[legcay] Use 'common.js' instead of 'common-legacy.js' in Tests.js
Instead of importing 'common-legacy.js' and using the self.Common
global, we import 'common.js' directly and stash the module in a
local variable.

Note that we can't use static imports as 'Tests.js' is loaded as a
classic script and loading 'Tests.js' as an ES module requires
non-trivial code changes in Chromium.

R=alexrudenko@chromium.org

Bug: 1442410
Change-Id: Ia5b47f2baae12ad7e1baf4a5bb8d2a34cabe3800