[search] Preserve search results in 'findFilesMatchingSearchRequest'
This CL changes the 'findFilesMatchingSearchRequest' return type from
a list of UISourceCodes to a Map keyed on UISourceCodes.
The ContentProviderBasedProject already does a prelminary search in
the backend as well as in the content and throws away the search result.

This CL is a first step to re-use the exact search results by
returning them. The "file based project" implementation
unconditionally omits the search result.

R=dsv@chromium.org

Bug: 1468875
Change-Id: Id46df819973e1caeb0dcd2b605012de35f60c145