[search] Fix response body search
This CL fixes NetworkSearchScope to hold "SearchConfig" correctly.
Currently we use "query()" for searching in the response body. This is
wrong as the "SearchConfig" does some parsing to the user entered
query.

We now mirror the same behavior as in file-based searching: For
requests to match, all the different sub-queries must produce matches.

Drive-by: NetworkRequest.searchInContent now always produces 'exact'
matches.

R=jarin@chromium.org

Bug: 1468875
Change-Id: I5810fc67d58b997131d1d185ea8b418d8daae60b